requirements.txt
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.2.2
openai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
huggingface-hub>=0.19.0
datasets>=2.14.0 